+++ pubtools/pulplib/_impl/fake/client.py
@@ -1,11 +1,12 @@
 import random
 import uuid
 import threading
+import hashlib
 
 from collections import namedtuple
 
 import six
-from more_executors.futures import f_return, f_return_error
+from more_executors.futures import f_return, f_return_error, f_flat_map
 
 from pubtools.pulplib import Page, PulpException, Criteria, Task
 from .. import compat_attr as attr
@@ -13,6 +14,7 @@ from .. import compat_attr as attr
 from .match import match_object
 
 Publish = namedtuple("Publish", ["repository", "tasks"])
+Upload = namedtuple("Upload", ["repository", "tasks", "name", "sha256"])
 
 
 class FakeClient(object):
@@ -28,6 +30,7 @@ class FakeClient(object):
     def __init__(self):
         self._repositories = []
         self._publish_history = []
+        self._upload_history = []
         self._lock = threading.RLock()
         self._uuidgen = random.Random()
         self._uuidgen.seed(0)
@@ -79,6 +82,52 @@ class FakeClient(object):
 
         return f_return(data[0])
 
+    def _do_upload_file(self, upload_id, file_obj, name):
+        # pylint: disable=unused-argument
+        is_file_obj = "close" in dir(file_obj)
+        if not is_file_obj:
+            file_obj = open(file_obj, "rb")
+
+        def do_next_upload(checksum, size):
+            data = file_obj.read(1024 * 1024)
+            if data:
+                checksum.update(data)
+                size += len(data)
+                return do_next_upload(checksum, size)
+            return f_return((checksum.hexdigest(), size))
+
+        out = f_flat_map(f_return(), lambda _: do_next_upload(hashlib.sha256(), 0))
+
+        if not is_file_obj:
+            out.add_done_callback(lambda _: file_obj.close())
+
+        return out
+
+    def _request_upload(self):
+        upload_request = {
+            "_href": "/pulp/api/v2/content/uploads/%s/" % self._next_request_id(),
+            "upload_id": "%s" % self._next_request_id(),
+        }
+
+        return f_return(upload_request)
+
+    def _do_import(self, repo_id, upload_id, unit_type_id, unit_key):
+        # pylint: disable=unused-argument
+        repo_f = self.get_repository(repo_id)
+        if repo_f.exception():
+            # Repo can't be found, let that exception propagate
+            return repo_f
+
+        repo = repo_f.result()
+
+        task = Task(id=self._next_task_id(), completed=True, succeeded=True)
+
+        self._upload_history.append(
+            Upload(repo, [task], unit_key["name"], unit_key["checksum"])
+        )
+
+        return f_return([task])
+
     def _delete_resource(self, resource_type, resource_id):
         if resource_type == "repositories":
             return self._delete_repository(resource_id)
@@ -128,3 +177,6 @@ class FakeClient(object):
         with self._lock:
             next_raw_id = self._uuidgen.randint(0, 2 ** 128)
         return str(uuid.UUID(int=next_raw_id))
+
+    def _next_request_id(self):
+        return self._next_task_id()