+++ pubtools/pulplib/_impl/fake/controller.py
@@ -79,3 +79,22 @@ class FakeController(object):
                 of this publish
         """
         return self.client._publish_history[:]
+
+    @property
+    def upload_history(self):
+        """A list of upload tasks triggered via this client.
+
+        Each element of this list is a named tuple with the following attributes,
+        in order:
+
+            ``repository``:
+                :class:`~pubtools.pulplib.Repository` for which upload was triggered
+            ``tasks``:
+                list of :class:`~pubtools.pulplib.Task` generated as a result
+                of this upload
+            ``name`` (str):
+                the remote path used
+            ``sha256`` (str):
+                checksum of the file uploaded
+        """
+        return self.client._upload_history[:]