+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- n/a
+### Added
+- A new API `FileRepository.upload_file` to upload a file to Pulp repository
 
 ## [1.1.0] - 2019-07-03